+++ service/service.go
@@ -914,7 +914,7 @@ func (s *Service) HandleHello(ownAddress, remoteAddress string, req *HelloReques
 					peer.Port = req.SlavePort
 					peer.DataDir = req.DataDir
 
-					peer.HasAgentFlag = boolFromRef(req.Agent, peer.HasAgentFlag)
+					peer.HasAgentFlag = utils.NotNilDefault(req.Agent, peer.HasAgentFlag)
 					peer.HasCoordinatorFlag = utils.NotNilDefault(req.Coordinator, peer.HasCoordinatorFlag)
 					peer.HasDBServerFlag = utils.NotNilDefault(req.DBServer, peer.HasDBServerFlag)
 					peer.HasResilientSingleFlag = boolFromRef(req.ResilientSingle, peer.HasResilientSingleFlag)
@@ -932,8 +932,9 @@ func (s *Service) HandleHello(ownAddress, remoteAddress string, req *HelloReques
 			portOffset := s.myPeers.GetFreePortOffset(slaveAddr, slavePort, s.cfg.AllPortOffsetsUnique)
 			s.log.Debug().Msgf("Set slave port offset to %d, got slaveAddr=%s, slavePort=%d", portOffset, slaveAddr, slavePort)
 
+			hasAgentFlag := !s.myPeers.HaveEnoughAgents()
 			servers := peerServers{
-				HasAgentFlag:           !s.myPeers.HaveEnoughAgents(),
+				HasAgentFlag:           &hasAgentFlag,
 				HasDBServerFlag:        nil,
 				HasCoordinatorFlag:     nil,
 				HasResilientSingleFlag: s.mode.IsActiveFailoverMode(),
@@ -941,7 +942,7 @@ func (s *Service) HandleHello(ownAddress, remoteAddress string, req *HelloReques
 				HasSyncWorkerFlag:      s.mode.SupportsArangoSync() && s.cfg.SyncEnabled,
 			}
 			if req.Agent != nil {
-				servers.HasAgentFlag = *req.Agent
+				servers.HasAgentFlag = req.Agent
 			}
 			if req.DBServer != nil && *req.DBServer != true {
 				servers.HasDBServerFlag = req.DBServer
@@ -1349,7 +1350,10 @@ func (s *Service) adjustClusterConfigForRelaunch(bsCfg BootstrapConfig) {
 	}
 
 	s.myPeers.ForEachPeer(func(p Peer) Peer {
-		p.peerServers = preparePeerServers(s.mode, bsCfg, s.cfg)
+		if bsCfg.ID == p.ID {
+			s.log.Debug().Msgf("Adjusting current memeber cluster config after restart (port: %d)", p.Port)
+			p.peerServers = preparePeerServers(s.mode, bsCfg, s.cfg)
+		}
 		return p
 	})
 }