+++ test/docker_util.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2017 ArangoDB GmbH, Cologne, Germany
+// Copyright 2017-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Ewout Prangsma
 //
 
 package test
@@ -51,9 +49,23 @@ func createDockerVolumes(t *testing.T, prefixes ...string) ([]string, func()) {
 }
 
 func createDockerVolume(t *testing.T, id string) {
-	c := Spawn(t, fmt.Sprintf("docker volume create %s", id))
+	c := Spawn(t, fmt.Sprintf("docker volume create %s --label starter-test=true", id))
 	defer c.Close()
 	c.Wait()
+}
+
+func removeDockerVolumesByLabel(t *testing.T, labelKeyValue string) {
+	ps := exec.Command("docker", "volume", "ls", "-q", "--filter", "label="+labelKeyValue)
+	list, err := ps.Output()
+	if err != nil {
+		t.Fatalf("docker ps failed: %s", describe(err))
+	}
+	ids := strings.Split(strings.TrimSpace(string(list)), "\n")
+	for _, id := range ids {
+		if id := strings.TrimSpace(id); id != "" {
+			removeDockerVolume(t, id)
+		}
+	}
 }
 
 func removeDockerVolume(t *testing.T, id string) {