+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [master](https://github.com/arangodb-helper/arangodb/tree/master) (N/A)
 - Bump Go version (1.22.2) and dependencies for CVE fixes
+- Fix issue with restart if member is using 'cluster.start-agent=false' feature
 
 ## [v0.18.4](https://github.com/arangodb-helper/arangodb/tree/0.18.4) (2024-03-11)
 - Fix upgrade with ArangoSync component enabled